change project name reader_tests => pdfras_reader_tests, and main from reader_test.c => pdfras_reader_tests.c
