+++ reader_test/reader_test.vcxproj
@@ -14,6 +14,7 @@
     <ProjectGuid>{34032D75-96A8-43EE-ADBF-F2B68AA4218D}</ProjectGuid>
     <Keyword>Win32Proj</Keyword>
     <RootNamespace>reader_test</RootNamespace>
+    <ProjectName>pdfras_reader_tests</ProjectName>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -81,7 +82,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\common\test_support.c" />
-    <ClCompile Include="reader_test.c" />
+    <ClCompile Include="pdfras_reader_tests.c" />
   </ItemGroup>
   <ItemGroup>
     <None Include="bad_trailer1.pdf" />
